hw-1_7/requirements.txt
fdb==2.0.2
python-dotenv==0.21.0
pandas==1.4.4
openpyxl==3.0.10
xlsxwriter==3.0.3
xlrd==2.0.1
requests==2.28.1
jinja2==3.1.2
